Move the `ClanMember` into the `component` package
